Move Api::V1::OrganizersController to Organizers module
